fix: adjust search url
https://mask.atlassian.net/browse/FW-2461
