+++ src/app/(normal)/search/[type]/layout.tsx
