+++ src/app/(normal)/search/[type]/loading.tsx
